testing code (#5)
* bitcoin API

* bcash API

* bitcoin API

* litecoin API

* ethereum API

* xrp API

* xlm API

* return pointer

* vet API

* ledger API

* fixed import packages

* testing code

* validate

* testing code

* kms, record APIs

+++ model/response/ltc/ltc_tx.go
@@ -42,7 +42,7 @@ type Tx struct {
 	 * @type {number}
 	 * @memberof LtcTx
 	 */
-	Height uint32
+	Height uint64
 	/**
 	 * Hash of the block this transaction belongs to.
 	 * @type {string}
